requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.11.0=py37haa95532_0
aiohttp=3.7.2=py37h2bbff1b_1
alabaster=0.7.12=py37_0
argh=0.26.2=py37_0
argon2-cffi=20.1.0=py37he774522_1
astor=0.8.1=py37_0
astroid=2.4.2=py37_0
async-timeout=3.0.1=py37_0
async_generator=1.10=py37h28b3542_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.4=py_0
babel=2.8.1=pyhd3eb1b0_0
backcall=0.2.0=py_0
bcrypt=3.2.0=py37he774522_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
certifi=2020.11.8=py37haa95532_0
cffi=1.14.3=py37hcd4344a_2
chardet=3.0.4=py37haa95532_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=py_0
cryptography=3.2.1=py37hcd4344a_1
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37he774522_0
dask-core=2.30.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20200713=py_0
docutils=0.16=py37_1
entrypoints=0.3=py37_0
flake8=3.8.4=py_0
freeglut=3.0.0=h6538335_5
freetype=2.10.4=hd328e21_0
future=0.18.2=py37_1
gast=0.2.2=py37_0
google-auth=1.23.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=67.1=h33f27b4_0
idna=2.10=py_0
imageio=2.9.0=py_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
imutils=0.5.3=py37h03978a9_3
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipynb-py-convert=0.4.6=pyh9f0ad1d_0
ipython=7.19.0=py37hd4e2768_0
ipython_genutils=0.2.0=py37_0
isort=5.6.4=py_0
jasper=2.0.14=hdc05fd1_1
jedi=0.17.1=py37_0
jinja2=2.11.2=py_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
keyring=21.4.0=py37_1
kiwisolver=1.3.0=py37hd77b12b_0
lazy-object-proxy=1.4.3=py37he774522_0
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libclang=10.0.1=default_hf44288c_1
liblapack=3.8.0=20_mkl
liblapacke=3.8.0=20_mkl
libopencv=4.5.0=py37_3
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.1.0=h56a325e_1
libwebp-base=1.1.0=he774522_3
lz4-c=1.9.2=hf4a77e7_3
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.3=py37haa95532_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib-base=3.3.2=py37hba9282a_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.2=256
mkl-service=2.3.0=py37h2bbff1b_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
msys2-conda-epoch=20160418=1
multidict=4.7.6=py37he774522_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
nest-asyncio=1.4.3=pyhd3eb1b0_0
networkx=2.5=py_0
notebook=6.1.4=py37_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
numpydoc=1.1.0=pyhd3eb1b0_1
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=4.5.0=py37_3
openssl=1.1.1h=he774522_0
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py37haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
pathtools=0.1.2=py_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=py37_1001
pillow=8.0.1=py37h4fa10fc_0
pip=20.2.4=py37haa95532_0
pluggy=0.13.1=py37_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37ha925a31_1
psutil=5.7.2=py37he774522_0
ptyprocess=0.6.0=pyhd3eb1b0_2
py-opencv=4.5.0=py37heaed05f_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=py_0
pygments=2.7.2=pyhd3eb1b0_0
pyjwt=1.7.1=py37_0
pylint=2.6.0=py37_0
pynacl=1.4.0=py37h62dcd97_1
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.12.3=py37h1834ac0_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyreadline=2.1=py37_1
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.35.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
pywavelets=1.1.1=py37he774522_2
pywin32=227=py37he774522_1
pywin32-ctypes=0.2.0=py37_1001
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=py37he774522_1
pyzmq=19.0.2=py37ha925a31_1
qdarkstyle=2.8.1=py_0
qt=5.12.9=hb2cf2c5_0
qtawesome=1.0.1=py_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rope=0.18.0=py_0
rsa=4.6=py_0
rtree=0.9.4=py37h21ff451_1
scikit-image=0.17.2=py37h1e1f486_0
scipy=1.5.2=py37h9439919_0
send2trash=1.5.0=py37_0
setuptools=50.3.1=py37haa95532_1
six=1.15.0=py37haa95532_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.2.2=py_0
sphinx=3.2.1=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.5=py37_0
spyder-kernels=1.9.4=py37_0
spyder-notebook=0.3.1=pyhd8ed1ab_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=eigen_py37hd727fc0_0
tensorflow-base=2.1.0=eigen_py37h49b2757_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
termcolor=1.1.0=py37_1
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
tifffile=2020.10.1=py37h8c2d366_2
tk=8.6.10=he774522_0
toml=0.10.1=py_0
toolz=0.11.1=py_0
tornado=6.0.4=py37he774522_1
traitlets=5.0.5=py_0
typed-ast=1.4.1=py37he774522_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
ujson=4.0.1=py37ha925a31_0
urllib3=1.25.11=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
watchdog=0.10.3=py37_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=0.14.1=py37_0
wheel=0.35.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.11.2=py37he774522_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.30.0=py_0
yarl=1.5.1=py37he774522_0
zeromq=4.3.2=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
